refactor(team): extract the lead relay capture stream hooks
TeamProvisioningStreamEvents.ts sits exactly on its frozen 1212-line cap and
is the file every new stream-json rule has to touch, so the lead relay
capture cannot grow another branch there without first moving out what it
already owns.

Move the capture's stream-side shape and its three stream transitions into
leadRelayCaptureStreamHooks.ts: the state interface the stream handler is
allowed to mutate, the synthetic-chunk test, the text join, the assistant
text append with its idle-resolve timer, and the terminal-result resolve.
The stream handler keeps the control flow and now calls the hooks; the
bodies are unchanged apart from taking the bounded-parts function as an
argument instead of reaching into ports.

Behaviour is identical: the same fields are set in the same order, the idle
timer is still cleared and rearmed on every captured text part, and a
terminal result still resolves the capture with the joined text. The point
is that the shape of a capture and the rules for advancing it now live in
one small file that a reviewer can read whole, instead of three places
inside a 1200-line stream switch.
